+++ lib/password_engine.dart
@@ -6,7 +6,6 @@
 library;
 
 export 'src/config/password_generator_config.dart';
-export 'src/constants/password_constants.dart';
 export 'src/generator/ipassword_generator.dart';
 export 'src/generator/password_generator.dart';
 export 'src/model/character_set_profile.dart';
@@ -16,5 +15,7 @@ export 'src/strategy/ipassword_generation_strategy.dart';
 export 'src/strategy/random_password_strategy.dart';
 export 'src/strength_estimator/ipassword_strength_estimator.dart';
 export 'src/strength_estimator/password_strength_estimator.dart';
+export 'src/validator/config_aware_password_validator.dart';
+export 'src/validator/iconfig_aware_password_validator.dart';
 export 'src/validator/ipassword_validator.dart';
 export 'src/validator/password_validator.dart';